Moved files to component folder
