+++ fontend/src/components/BlueGradientBar.js
