--- fontend/src/components/edit-todo.component.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-export default class EditTodo extends Component {
-        render()
-        {
-            return(
-                <div>
-                    <p>Welcome to edit Todo Component!</p>
-                </div>
-            )
-        }
-
-}
