--- fontend/src/components/todos-list.component.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-export default class TodosList extends Component {
-        render()
-        {
-            return(
-                <div>
-                    <p>Welcome to todos List Component!</p>
-                </div>
-            )
-        }
-
-}
